crispr_env/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "crispr-env"
version = "1.0.0"
description = "CRISPR Guide RNA Design Environment for OpenEnv hackathon"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "VidhaanKhare16" }]
dependencies = [
    "openenv-core>=0.2.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx"]

[tool.setuptools.packages.find]
where = ["."]
include = ["crispr_env*"]

[tool.uv]
dev-dependencies = ["pytest>=8.0", "httpx>=0.27"]

[project.scripts]
server = "uvicorn crispr_env.server.app:app --host 0.0.0.0 --port 8000 --reload"
